feat: start adding support for aarch64/arm NEON
