+++ src/models/common.rs
@@ -7,7 +7,6 @@ use async_std::stream::Cloned;
 use dashmap::DashMap;
 use futures::future::{join_all, BoxFuture, FutureExt};
 use sha2::{Digest, Sha256};
-use std::arch::x86_64::*;
 use std::collections::hash_map::DefaultHasher;
 use std::collections::HashSet;
 use std::fmt;
@@ -16,6 +15,10 @@ use std::sync::Arc;
 use thiserror::Error;
 use tokio::task;
 
+#[cfg(target_arch = "x86_64")]
+use std::arch::x86_64::*;
+
+#[cfg(target_arch = "x86_64")]
 pub fn dot_product_u8_avx2_fma(a: &[u8], b: &[u8]) -> u64 {
     assert_eq!(a.len(), b.len());
